id: 64, tempo 3 minuti
La risoluzione presente in questo commit risolve le issues con id 64 , in quanto la issue richiedeva di ritornare un double. il tempo stimato è di 3 minuti a issue

+++ src/main/java/com/github/mreutegg/laszip4j/lastools/Laszip.java
@@ -759,7 +759,7 @@ public class Laszip {
     }
 
     private static double taketime() {
-        return System.currentTimeMillis() / 1000;
+        return (double) System.currentTimeMillis() / 1000;
     }
 
     private static void usage() {
